Rename package "src" to "sampletester"
- rename sampletester.py to cli.py
