+++ examples/convention-cloud/run.sh
@@ -13,4 +13,4 @@
 # limitations under the License.
 
 echo "*** " $(pwd)
-python3 -m src.sampletester ${FLAGS:---xunit $(mktemp --suffix=.xml --tmpdir sampletester.xunit.XXXXX) -v detailed} --convention=cloud examples/convention-cloud/language.test.yaml examples/mock-samples/googleapis 
+python3 -m sampletester.cli ${FLAGS:---xunit $(mktemp --suffix=.xml --tmpdir sampletester.xunit.XXXXX) -v detailed} --convention=cloud examples/convention-cloud/language.test.yaml examples/mock-samples/googleapis
