+++ sampletester/sample_manifest.py
